Crate crust

crust

Reliable peer-to-peer network connections in Rust with NAT traversal.

Structs

ConnectionInfoResult
The result of a Service::prepare_contact_info call.

Endpoint
Enum representing endpoint of supported protocols

OurConnectionInfo
Contact info generated by a call to Service::prepare_contact_info.

PeerId
An identifier of a peer node.

Service
A structure representing a connection manager.

SocketAddr
Wrapper around std::net::SocketAddr to enable it to encoded and decoded.

StaticContactInfo
This struct contains information needed to Bootstrap and for echo-server services

TheirConnectionInfo
Contact info used to connect to another peer.

Enums

Error
Error types.

Event
Enum representing different events that will be sent over the asynchronous channel to the user of this module.

Protocol
Enum representing supported transport protocols

Functions

write_config_file
Writes a Crust config file for use by tests and examples.

Type Definitions

CrustEventSender
Crust Observers will be informed of crust events on this

Implementations