Module crust::connection_manager

Module implementing the ConnectionManager which provides an interface to manage peer-to-peer connections.

Structs

ConnectionManager
A structure representing a connection manager

Enums

Event
Enum representing different events that will be sent over the asynchronous channel to the user of this module.

Type Definitions

Bytes
Type used to represent serialised data in a message.