Enum crust::connection_manager::Event

pub enum Event {
    NewMessage(Endpoint, Bytes),
    NewConnection(Endpoint),
    LostConnection(Endpoint),
}

Enum representing different events that will be sent over the asynchronous channel to the user of this module.

Variants

NewMessage

Invoked when a new message is received. Passes the peer's endpoint and the message.

NewConnection

Invoked when a new connection to a peer is established. Passes the peer's endpoint.

LostConnection

Invoked when a connection to a peer is lost. Passes the peer's endpoint.