Struct crust::connection_manager::ConnectionManager

pub struct ConnectionManager {
    // some fields omitted
}

A structure representing a connection manager

Methods

impl ConnectionManager

fn new(event_pipe: IoSender<Event>) -> ConnectionManager

Constructs a connection manager. User needs to create an asynchronous channel, and provide the sender half to this method. Receiver half will recieve all the events Event from this library.

fn start_listening(&self, hint: Vec<Port>) -> IoResult<Vec<Endpoint>>

Starts listening on all supported protocols. Specified hint will be tried first, if it fails to start on these, it defaults to random / OS provided endpoints for each supported protocol. The actual endpoints used will be returned on which it started listening for each protocol.

fn bootstrap(&self, bootstrap_list: Option<Vec<Endpoint>>) -> IoResult<Endpoint>

This method tries to connect (bootstrap to exisiting network) to the default or provided list of bootstrap nodes. If the bootstrap list is Some, the method will try to connect to all of the endpoints specified in bootstrap_list. It will return once connection with any of the endpoint is established with Ok(Endpoint) and it will drop all other ongoing attempts. Returns Err if if fails to connect to any of the endpoints specified. If bootstrap_list is None, it will use default methods to bootstrap to the existing network. Default methods includes beacon system for finding nodes on a local network and bootstrap handler which will attempt to reconnect to any previous "direct connected" nodes. In both cases, this method blocks until it gets one successful connection or all the endpoints are tried and failed.

fn connect(&self, endpoints: Vec<Endpoint>)

Opens a connection to a remote peer. endpoints is a vector of addresses of the remote peer. All the endpoints will be tried. As soon as one of the connection is established, it will drop all other ongoing attempt. On success Event::NewConnection with connected Endpoint will be sent to the event channel. On failure to connect to any of the provided endpoints, Event::FailedToConnect will be sent to the event channel. Failed attempts are not notified back up to the caller. If the caller wants to know of a failed attempt, it must maintain a record of the attempt itself which times out if a corresponding Event::NewConnection isn't received For details on handling of connect in different protocol refer https://github.com/dirvine/crust/blob/master/docs/connect.md

fn send(&self, endpoint: Endpoint, message: Bytes) -> IoResult<()>

Sends a message to specified address (endpoint). Returns Ok(()) if the sending might succeed, and returns an Err if the address is not connected. Return value of Ok does not mean that the data will be received. It is possible for the corresponding connection to hang up immediately after this function returns Ok.

fn drop_node(&self, endpoint: Endpoint)

Closes connection with the specified endpoint

fn get_stored_bootstrap_endpoints(&self) -> Vec<Endpoint>