Enum crust::Event

pub enum Event {
    BootstrapAccept(PeerId),
    BootstrapConnect(PeerId, SocketAddr),
    BootstrapFailed,
    ListenerStarted(u16),
    ListenerFailed,
    ConnectionInfoPrepared(ConnectionInfoResult),
    ConnectSuccess(PeerId),
    ConnectFailure(PeerId),
    LostPeer(PeerId),
    NewMessage(PeerId, Vec<u8>),
    WriteMsgSizeProhibitive(PeerId, Vec<u8>),
}

Enum representing different events that will be sent over the asynchronous channel to the user of this module.

Variants

BootstrapAccept

Invoked when a bootstrap peer connects to us

BootstrapConnect

Invoked when we bootstrap to a new peer.

BootstrapFailed

Invoked when we failed to connect to all bootstrap contacts.

ListenerStarted

Invoked when we are ready to listen for incomming connection. Contains the listening port.

ListenerFailed

Invoked when listener failed to start.

ConnectionInfoPrepared

Invoked as a result to the call of Service::prepare_contact_info.

ConnectSuccess

Invoked when connection to a new peer has been established.

ConnectFailure

Invoked when connection to a new peer has failed.

LostPeer

Invoked when a peer disconnects or can no longer be contacted.

NewMessage

Invoked when a new message is received. Passes the message.

WriteMsgSizeProhibitive

Invoked when trying to sending a too large data.