Struct crust::file_handler::FileHandler

pub struct FileHandler {
    // some fields omitted
}

Struct for reading and writing config files.

Thread- and Process-Safety

Since all instances of FileHandler initialised with the same value for name within a single process will likely refer to the same file on disk, it is not safe to access any such instance concurrently with that same or any other such instance (with the exception of the path() function which is the only non-mutating member function).

For instances initialised with different values for name, it is safe to access separate instances concurrently.

It is possibly unsafe to call write_file() concurrently with a different process calling read_file() or write_file() where both processes have the same name and their instances of FileHandler are using the same name, since these may be accessing the same file on disk. However, it is safe to call read_file() concurrently across multiple such processes, since this function doesn't modify the file.

Perhaps the easiest way to make multi-process access safe is to ensure each process is a single execution of a binary, and that each binary is located in a directory which is mutually-exclusive to all other such binaries, and that each config file to be managed by FileHandler is placed in each binary's current_bin_dir(). In this way, each process should be the only one accessing that file.

Methods

impl FileHandler

fn new(name: PathBuf) -> FileHandler

Constructor taking the required file name (not the full path)

fn read_file<Contents: Decodable>(&mut self) -> Result<Contents, Error>

Reads the file and returns the JSON-decoded contents or an error.

It tries to read from the following locations in this order (see also an example config file flowchart):

The location of the most recent successful read or write attempt

current_bin_dir()

user_app_dir()

system_cache_dir()

See Thread- and Process-Safety for notes on thread- and process-safety.

NOTE

If a file (or directory) exists at the expected location when an attempt is made to read it, and reading or decoding fails, this function WILL TERMINATE THE APPLICATION. This means that the only cause for an error to be returned from this function is the non-existence of the file in all attempted locations.

fn write_file<Contents: Encodable>(&mut self, contents: &Contents) -> Result<(), Error>

JSON-encodes then writes contents to the file. Creates the file if it doesn't already exist.

If contents fails to encode or the file cannot be written, an error is returned. The process is (see also an example config file flowchart):

If the file has previously been read (i.e. path() is Some(...)), it tries to write the contents to this path. If this fails, it jumps to step 3.

It tries to create and write the file in system_cache_dir(). It will not try and create this directory if it doesn't exist.

It tries to create and write the file in user_app_dir(). It will try to create this directory and any parent components if they don't exist.

See Thread- and Process-Safety for notes on thread- and process-safety.

fn path(&self) -> &Option<PathBuf>

Get the full path to the file.

If no calls to read_file() or write_file() have been made, or the last such attempt failed, then this will return None.

See Thread- and Process-Safety for notes on thread- and process-safety.