Struct crust::file_handler::ScopedUserAppDirRemover

pub struct ScopedUserAppDirRemover;

RAII object which removes the user_app_dir() when an instance is dropped.

Since the user_app_dir is frequently created by tests or examples which use Crust, this is a convenience object which tries to remove the directory when it is destroyed.

Examples

{
    let _cleaner = ::crust::ScopedUserAppDirRemover;
    let mut file_handler =
        ::crust::FileHandler::new(::std::path::Path::new("test.json").to_path_buf());
    // User app dir is possibly created by this call.
    let _ = file_handler.write_file(&111u64);
}
// User app dir is now removed since '_cleaner' has gone out of scope.