Function crust::write_config_file

pub fn write_config_file(hard_coded_endpoints: Option<Vec<StaticContactInfo>>) -> Result<PathBuf, Error>

Writes a Crust config file for use by tests and examples.

The file is written to the current_bin_dir() with the appropriate file name.

N.B. This method should only be used as a utility for test and examples. In normal use cases, this file should be created by the installer for the dependent application.