Struct crust::service::Service

pub struct Service {
    // some fields omitted
}

A structure representing a connection manager.

This abstraction has a hidden dependency on a config file. Refer to the docs for FileHandler and an example config file flowchart for more information.

Methods

impl Service

fn new(event_sender: CrustEventSender) -> Result<Service>

Constructs a service. User needs to create an asynchronous channel, and provide the sender half to this method. Receiver will receive all Events from this library.

fn new_inactive(event_sender: CrustEventSender) -> Result<Service>

Construct a service. As with the Service::new function, but will not implicitly start any network activity. This constructor is intended only for testing purposes.

fn start_beacon(&mut self, udp_port: u16) -> Result<u16>

Start the beaconing on port udp_port. If port number is 0, the OS will pick one randomly. The actual port used will be returned.

This function MUST NOT be called more than once. Currently crust has a limit of listenning at most once per process.

fn start_accepting(&mut self, port: Port) -> Result<Endpoint>

Starts accepting on a given port. If port number is 0, the OS will pick one randomly. The actual port used will be returned.

fn bootstrap(&mut self, token: u32, beacon_port: Option<u16>)

This method tries to connect (bootstrap to existing network) to the default or provided override list of bootstrap nodes (via config file named .config).

If override_default_bootstrap_methods is not set in the config file, it will attempt to read a local cached file named .bootstrap.cache to populate the list endpoints to use for bootstrapping. It will also try hard_coded_contacts from config file. In addition, it will try to use the beacon port (provided via config file) to connect to a peer on the same LAN. For more details on bootstrap cache file refer https://github.com/maidsafe/crust/blob/master/docs/bootstrap.md

If override_default_bootstrap_methods is set in config file, it will only try to connect to the endpoints in the override list (hard_coded_contacts). All connections (if any) will be dropped before bootstrap attempt is made. This method returns immediately after dropping any active connections.endpoints New bootstrap connections will be notified by NewBootstrapConnection event. Its upper layer's responsibility to maintain or drop these connections.

fn bootstrap_with_blacklist(&mut self, token: u32, beacon_port: Option<u16>, blacklist: &[Endpoint])

Same as bootstrap, but allows to specify a blacklist of endpoints this service should never connect to.

fn stop_bootstrap(&mut self)

Stop the bootstraping procedure

fn remove_bootstrap_contact(&mut self, endpoint: Endpoint)

Remove endpoint from the bootstrap cache.

fn stop(&mut self)

This should be called before destroying an instance of a Service to allow the listener threads to join. Once called, the Service should be destroyed.

fn connect(&self, token: u32, endpoints: Vec<Endpoint>)

Opens a connection to a remote peer. endpoints is a vector of addresses of the remote peer. All the endpoints will be tried. As soon as a connection is established, it will drop all other ongoing attempts. On success Event::NewConnection with connected Endpoint will be sent to the event channel. On failure, nothing is reported. Failed attempts are not notified back up to the caller. If the caller wants to know of a failed attempt, it must maintain a record of the attempt itself which times out if a corresponding Event::NewConnection isn't received. See also Process for Connecting for details on handling of connect in different protocols.

fn rendezvous_connect(&self, udp_socket: UdpSocket, token: u32, public_endpoint: Endpoint)

Opens a connection to a remote peer. public_endpoint is the endpoint of the remote peer. udp_socket is a socket whose public address will be used by the other peer.

A rendezvous connection setup is different to the traditional BSD socket setup in which there is no client or server side. Both ends create a socket and send somehow its public address to the other peer. Once both ends know each other address, both must call this function passing the socket which possess the address used by the other peer and passing the other peer's address.

Only UDP-based protocols are supported. This means that you must use a uTP endpoint or nothing will happen.

On success Event::OnRendezvousConnect with connected Endpoint will be sent to the event channel. On failure, nothing is reported. Failed attempts are not notified back up to the caller. If the caller wants to know of a failed attempt, it must maintain a record of the attempt itself which times out if a corresponding Event::OnRendezvousConnection isn't received. See also Process for Connecting for details on handling of connect in different protocols.

fn send(&self, connection: Connection, message: Bytes)

Sends a message over a specified connection.

fn drop_node(&self, connection: Connection)

Closes a connection.

fn get_external_endpoints(&self)

Initiates uPnP port mapping of the currently used accepting endpoints. On success ExternalEndpoint event is generated containg our external endpoints.

fn get_mapped_udp_socket(&self, result_token: u32)

Lookup a mapped udp socket based on result_token

fn udp_punch_hole(&self, result_token: u32, udp_socket: UdpSocket, secret: Option<[u8; 4]>, peer_addr: SocketAddr)

Udp hole punching process