Struct crust::Connection

pub struct Connection {
    // some fields omitted
}

An open connection that can be used to send messages to a peer.

Messages from the peer are received as Crust events, together with the peer's public key.

The connection is closed when this is dropped.

Methods

impl Connection

fn send(&mut self, data: &[u8]) -> Result<()>

Send the data to a peer via this connection.

fn is_closed(&self) -> bool

Returns whether this connection has been closed.