Struct crust::OurContactInfo

pub struct OurContactInfo {
    pub socket: UdpSocket,
    pub secret: Option<[u8; 4]>,
    pub static_addrs: Vec<Endpoint>,
    pub rendezvous_addrs: Vec<SocketAddr>,
}

Contact info generated by a call to Service::prepare_contact_info.

Fields

socket

The mapped udp socket that may be used for this connection.

secret

Secret data used for rendezvous connect.

static_addrs

Our tcp listening addresses.

rendezvous_addrs

The mapped addresses of our udp socket.

Methods

impl OurContactInfo

fn make_their_info(&self) -> TheirContactInfo

Takes an OurContactInfo and turns it into a TheirContactInfo. This TheirContactInfo can then be sent out-of-band to another peer so that they can use it to connect to us.