Struct crust::StaticContactInfo

pub struct StaticContactInfo {
    pub tcp_acceptors: Vec<SocketAddr>,
    pub utp_custom_listeners: Vec<SocketAddr>,
    pub mapper_servers: Vec<SocketAddr>,
}

This struct contains information needed to Bootstrap and for echo-server services

Fields

tcp_acceptors

This will contain both local and global addresses. Local addresses will be useful on LAN when someone wants to bootstrap off us and we haven't yet obtained our global address. In that case the list will contain only the local addresses that the process calling seek_peers() will get and use.

utp_custom_listeners

This will contain only global addresses. Local addresses will be useful on LAN when someone has these entries in bootstrap.cache file and crust was directed not to seek_peers() during start (seek_peers_on_port = None) and TCP was not allowed on that Lan. In this case the peer would ask for a new UDP socket address to connect to from one of the following listeners. Since this is a rare edge case we do not cater to it and have only global addresses populated for UDP.

mapper_servers

Mapper server addresses