Crate crust

crust

Reliable peer-to-peer network connections in Rust with NAT traversal.

Reexports

pub use service::Service;

pub use error::Error;

pub use file_handler::{FileHandler, current_bin_dir, user_app_dir, system_cache_dir, exe_file_stem, ScopedUserAppDirRemover};

Modules

error
Defines errors.

file_handler
Provides a struct and free functions for working with config files.

service
Module implementing the Service which provides an interface to manage peer-to-peer connections.

Structs

Connection
Information hold for the connection between a pair of nodes

SocketAddrW
Utility struct of SocketAddr for hole punching

Enums

Endpoint
Enum representing endpoint of supported protocols

Event
Enum representing different events that will be sent over the asynchronous channel to the user of this module.

Port
Enum representing port of supported protocols

Protocol
Enum representing supported transport protocols

Functions

getifaddrs
For non-Windows operating system, use this function to get address

ifaddrs_if_unspecified
This function should really take IpAddr as an argument but it is used outside of this library and IpAddr is currently considered experimental.

write_config_file
Writes a Crust config file for use by tests and examples.