Crate crust

crust

Reliable peer-to-peer network connections in Rust with NAT traversal.

Reexports

pub use connection_manager::{Event, ConnectionManager};

pub use error::Error;

pub use file_handler::{FileHandler, current_bin_dir, user_app_dir, system_cache_dir, exe_file_stem, ScopedUserAppDirRemover};

Modules

connection_manager
Module implementing the ConnectionManager which provides an interface to manage peer-to-peer connections.

error
Defines errors.

file_handler
Provides a struct and free functions for working with config files.

Enums

Endpoint
Enum representing endpoint of supported protocols

Port
Enum representing port of supported protocols

Functions

write_config_file
Writes a Crust config file for use by tests and examples.