Struct crust::connection_manager::ConnectionManager

pub struct ConnectionManager {
    // some fields omitted
}

A structure representing a connection manager

This abstraction has a hidden dependency on per-executable config file. Refer to https://github.com/maidsafe/crust/blob/master/docs/configuration_path.md for more information.

Methods

impl ConnectionManager

fn new(event_pipe: Sender<Event>) -> ConnectionManager

Constructs a connection manager. User needs to create an asynchronous channel, and provide the sender half to this method. Receiver will receive all Events from this library.

fn start_accepting(&mut self, hint: Vec<Port>) -> Result<Vec<Port>>

Starts listening on all supported protocols. Ports in hint are tried first. On failure to listen on none of hint an OS randomly chosen port will be used for each supported protocol. The actual port used will be returned on which it started listening for each protocol.

fn bootstrap(&mut self, max_successful_bootstrap_connection: usize)

This method tries to connect (bootstrap to existing network) to the default or provided override list of bootstrap nodes (via config file named .config).

If override_default_bootstrap_methods is not set in the config file, it will attempt to read a local cached file named .bootstrap.cache to populate the list endpoints to use for bootstrapping. It will also try hard_coded_contacts from config file. In addition, it will try to use the beacon port (provided via config file) to connect to a peer on the same LAN. For more details on bootstrap cache file refer https://github.com/maidsafe/crust/blob/master/docs/bootstrap.md

If override_default_bootstrap_methods is set in config file, it will only try to connect to the endpoints in the override list (hard_coded_contacts). All connections (if any) will be dropped before bootstrap attempt is made. This method returns immediately after dropping any active connections.endpoints New bootstrap connections will be notified by NewBootstrapConnection event. Its upper layer's responsibility to maintain or drop these connections. Maximum of max_successful_bootstrap_connection bootstrap connections will be made and further connection attempts will stop. It will reiterate the list of all endpoints until it gets at least one connection.

fn stop(&mut self)

This should be called before destroying an instance of a ConnectionManager to allow the listener threads to join. Once called, the ConnectionManager should be destroyed.

fn connect(&self, endpoints: Vec<Endpoint>)

Opens a connection to a remote peer. endpoints is a vector of addresses of the remote peer. All the endpoints will be tried. As soon as a connection is established, it will drop all other ongoing attempts. On success Event::NewConnection with connected Endpoint will be sent to the event channel. On failure, nothing is reported. Failed attempts are not notified back up to the caller. If the caller wants to know of a failed attempt, it must maintain a record of the attempt itself which times out if a corresponding Event::NewConnection isn't received For details on handling of connect in different protocol refer https://github.com/dirvine/crust/blob/master/docs/connect.md

fn send(&self, endpoint: Endpoint, message: Bytes) -> Result<()>

Sends a message to specified address (endpoint). Returns Ok(()) if the sending might succeed, and returns an Err if the address is not connected. Return value of Ok does not mean that the data will be received. It is possible for the corresponding connection to hang up immediately after this function returns Ok.

fn drop_node(&self, endpoint: Endpoint)

Closes connection with the specified endpoint.

fn get_own_endpoints(&self) -> Vec<Endpoint>

Return the endpoints other peers can use to connect to. External address are obtained through UPnP IGD.