Struct crust::connection_manager::ConnectionManager

pub struct ConnectionManager {
    // some fields omitted
}

A structure representing a connection manager

Methods

impl ConnectionManager

fn new(event_pipe: Sender<Event>) -> ConnectionManager

Constructs a connection manager. User needs to create an asynchronous channel, and provide the sender half to this method. Receiver will receive all Events from this library.

fn start_listening(&mut self, hint: Vec<Port>, beacon_port: Option<u16>) -> Result<(Vec<Port>, Option<u16>)>

Starts listening on all supported protocols. Specified hint will be tried first. If it fails to start on these, it defaults to random / OS provided endpoints for each supported protocol. The actual endpoints used will be returned on which it started listening for each protocol. if beacon port == 0 => a random port is taken and returned by beacon if beacon port != 0 => an attempt to get the port is made by beacon and the callee will be informed of the attempt if beacon port == None => 5483 is tried if beacon succeeds in starting the udp listener, the coresponding port is returned

fn start_listening2(&mut self, hint: Vec<Port>, beacon_port: Option<u16>) -> Result<(Vec<Endpoint>, Option<u16>)>

For API compatibilty, return a vector of listening endpoints

fn bootstrap(&self, bootstrap_list: Option<Vec<Endpoint>>, beacon_port: Option<u16>) -> Result<Endpoint>

This method tries to connect (bootstrap to exisiting network) to the default or provided list of bootstrap nodes.

If bootstrap_list is None, it will attempt to read a local cached file to populate the list. It will then try to connect to all of the endpoints in the list. It will return once a connection with any of the endpoints is established with Ok(Endpoint) and it will drop all other ongoing attempts. If this fails and beacon_port is Some, it will try to use the beacon port to connect to a peer on the same LAN. For more details on bootstrap cache file refer https://github.com/maidsafe/crust/blob/master/docs/bootstrap.md

If bootstrap_list is Some, it will try to connect to all of the endpoints in the list. It will return once a connection with any of the endpoints is established with Ok(Endpoint) and it will drop all other ongoing attempts. Note that beacon_port has no effect if bootstrap_list is Some; i.e. passing an explicit list ensures we only get connected to one of the nodes on the list - we don't fall back to use the beacon protocol.

It will return Err if it fails to connect to any peer.

fn stop(&mut self)

This should be called before destroying an instance of a ConnectionManager to allow the listener threads to join. Once called, the ConnectionManager should be destroyed.

fn connect(&self, endpoints: Vec<Endpoint>)

Opens a connection to a remote peer. endpoints is a vector of addresses of the remote peer. All the endpoints will be tried. As soon as a connection is established, it will drop all other ongoing attempts. On success Event::NewConnection with connected Endpoint will be sent to the event channel. On failure, nothing is reported. Failed attempts are not notified back up to the caller. If the caller wants to know of a failed attempt, it must maintain a record of the attempt itself which times out if a corresponding Event::NewConnection isn't received For details on handling of connect in different protocol refer https://github.com/dirvine/crust/blob/master/docs/connect.md

fn send(&self, endpoint: Endpoint, message: Bytes) -> Result<()>

Sends a message to specified address (endpoint). Returns Ok(()) if the sending might succeed, and returns an Err if the address is not connected. Return value of Ok does not mean that the data will be received. It is possible for the corresponding connection to hang up immediately after this function returns Ok.

fn drop_node(&self, endpoint: Endpoint)

Closes connection with the specified endpoint.