Struct crust::connection_manager::ConnectionManager

pub struct ConnectionManager<Id: Hash + Eq> {
    // some fields omitted
}

Methods

impl<Id> ConnectionManager<Id> where Id: Hash + Eq + Send + 'static + Clone + Encodable + Decodable + Debug

fn new(our_id: Id, event_pipe: IoSender<Event<Id>>) -> ConnectionManager<Id>

fn start_accepting(&self) -> IoResult<u16>

fn connect(&self, endpoint: SocketAddr, msg: Bytes) -> IoResult<()>

fn send(&self, message: Bytes, id: Id) -> IoResult<()>

Sends a message to address. Returns Ok(()) if the sending might succeed, and returns an Err if the address is not connected. Return value of Ok does not mean that the data will be received. It is possible for the corresponding connection to hang up immediately after this function returns Ok.

fn drop_node(&self, id: Id) -> IoResult<()>

fn id(&self) -> Id

fn stop(&self)