Enum crust::Event

pub enum Event {
    NewMessage(PeerId, Vec<u8>),
    BootstrapConnect(PeerId),
    BootstrapAccept(PeerId),
    NewPeer(Result<()>, PeerId),
    LostPeer(PeerId),
    BootstrapFinished,
    ConnectionInfoPrepared(ConnectionInfoResult),
}

Enum representing different events that will be sent over the asynchronous channel to the user of this module.

Variants

NewMessage

Invoked when a new message is received. Passes the message.

BootstrapConnect

Invoked when we get a bootstrap connection to a new peer.

BootstrapAccept

Invoked when a bootstrap peer connects to us

NewPeer

Invoked when a connection to a new peer is established.

LostPeer

Invoked when a peer is lost.

BootstrapFinished

Raised once the list of bootstrap contacts is exhausted.

ConnectionInfoPrepared

Invoked as a result to the call of Service::prepare_contact_info.