Enum crust::Event

pub enum Event {
    NewMessage(Connection, Vec<u8>),
    OnConnect(Result<(Endpoint, Connection)>, u32),
    OnRendezvousConnect(Result<(Endpoint, Connection)>, u32),
    OnAccept(Endpoint, Connection),
    LostConnection(Connection),
    BootstrapFinished,
    ExternalEndpoints(Vec<Endpoint>),
    OnUdpSocketMapped(MappedUdpSocket),
    OnHolePunched(HolePunchResult),
}

Enum representing different events that will be sent over the asynchronous channel to the user of this module.

Variants

NewMessage

Invoked when a new message is received. Passes the peer's endpoint and the message.

OnConnect

Invoked when the new connection request finishes. Passes the peer's endpoint and the token used in the request.

OnRendezvousConnect

Invoked when the new rendezvous connection request finishes. Passes the peer's endpoint and the token used in the request.

OnAccept

Invoked when a new connection is accepted. Passes the peer's endpoint.

LostConnection

Invoked when a connection to a peer is lost. Passes the peer's endpoint.

BootstrapFinished

Invoked when a new bootstrap connection to a peer is established. Passes the peer's endpoint.

ExternalEndpoints

Invoked when a new bootstrap connection to a peer is established. Passes the peer's endpoint.

OnUdpSocketMapped

Invoked as a result of the call to Service::get_mapped_udp_socket.

OnHolePunched

Invoked as a result of the call to Service::udp_punch_hole.