Struct crust::file_handler::FileHandler

pub struct FileHandler {
    // some fields omitted
}

Struct for reading and writing to config files

Methods

impl FileHandler

fn new(name: PathBuf) -> FileHandler

Constructor taking the required file name (not the full path)

fn read_file<Contents: Decodable>(&mut self) -> Result<Contents, Error>

Reads the file and returns the JSON-decoded contents or an error.

It tries to read from the following locations in this order:

The location of the most recent successful read or write attempt

current_bin_dir()

user_app_dir()

system_cache_dir()

fn write_file<Contents: Encodable>(&mut self, contents: &Contents) -> Result<(), Error>

JSON-encodes then writes contents to the file. Creates the file if it doesn't already exist.

If contents fails to encode or the file cannot be written, an error is returned. The process is:

If the file has previously been read (i.e. path() is Some(...)), it tries to write the contents to this path. If this fails, it jumps to step 3.

It tries to create and write the file in system_cache_dir(). It will not try and create this directory if it doesn't exist.

It tries to create and write the file in user_app_dir(). It will try to create this directory and any parent components if they don't exist.

fn path(&self) -> &Option<PathBuf>

Get the full path to the file.

If no calls to read_file() or write_file() have been made, or the last such attempt failed, then this will return None.