Struct crust::service::Service

pub struct Service {
    // some fields omitted
}

A structure representing a connection manager.

This abstraction has a hidden dependency on a config file. Refer to the docs for FileHandler and an example config file flowchart for more information.

Methods

impl Service

fn new(event_sender: Sender<Event>) -> Result<Service>

Constructs a service. User needs to create an asynchronous channel, and provide the sender half to this method. Receiver will receive all Events from this library.

fn new_inactive(event_sender: Sender<Event>) -> Result<Service>

Construct a service. As with the Service::new function, but will not implicitly start any network activity. This construtor is intended only for testing purposes.

fn start_default_acceptors(&mut self) -> Vec<Result<Endpoint>>

Start accepting on ports defined in the config file. Returns a vector of Ok(endpoint) for endpoints where the accetation started successfully.

fn start_accepting(&mut self, port: Port) -> Result<Endpoint>

Starts accepting on a given port. If port number is 0, the OS will pick one randomly. The actual port used will be returned.

fn bootstrap(&mut self)

This method tries to connect (bootstrap to existing network) to the default or provided override list of bootstrap nodes (via config file named .config).

If override_default_bootstrap_methods is not set in the config file, it will attempt to read a local cached file named .bootstrap.cache to populate the list endpoints to use for bootstrapping. It will also try hard_coded_contacts from config file. In addition, it will try to use the beacon port (provided via config file) to connect to a peer on the same LAN. For more details on bootstrap cache file refer https://github.com/maidsafe/crust/blob/master/docs/bootstrap.md

If override_default_bootstrap_methods is set in config file, it will only try to connect to the endpoints in the override list (hard_coded_contacts). All connections (if any) will be dropped before bootstrap attempt is made. This method returns immediately after dropping any active connections.endpoints New bootstrap connections will be notified by NewBootstrapConnection event. Its upper layer's responsibility to maintain or drop these connections.

fn stop_bootstrap(&mut self)

fn stop(&mut self)

This should be called before destroying an instance of a Service to allow the listener threads to join. Once called, the Service should be destroyed.

fn connect(&self, endpoints: Vec<Endpoint>)

Opens a connection to a remote peer. endpoints is a vector of addresses of the remote peer. All the endpoints will be tried. As soon as a connection is established, it will drop all other ongoing attempts. On success Event::NewConnection with connected Endpoint will be sent to the event channel. On failure, nothing is reported. Failed attempts are not notified back up to the caller. If the caller wants to know of a failed attempt, it must maintain a record of the attempt itself which times out if a corresponding Event::NewConnection isn't received. See also Process for Connecting for details on handling of connect in different protocols.

fn send(&self, connection: Connection, message: Bytes)

Sends a message over a specified connection.

fn drop_node(&self, connection: Connection)

Closes a connection.

fn get_external_endpoints(&self)

Initiates uPnP port mapping of the currently used accepting endpoints. On success ExternalEndpoint event is generated containg our external endpoints.