Struct crust::Service

pub struct Service {
    // some fields omitted
}

A structure representing a connection manager.

This abstraction has a hidden dependency on a config file. Refer to the docs for FileHandler and an example config file flowchart for more information.

Methods

impl Service

fn new(event_tx: CrustEventSender, service_discovery_port: u16) -> Result<Service, Error>

Constructs a service. User needs to create an asynchronous channel, and provide the sender half to this method. Receiver will receive all Events from this library.

fn stop_bootstrap(&mut self)

Stop the bootstraping procedure

fn set_listen_for_peers(&self, listen: bool) -> bool

Enable or Disable listening to peers trying to find us. The return value indicates successful registration of the request.

fn get_ordered_helping_nodes(&self) -> Vec<SocketAddr>

Get the hole punch servers addresses of nodes that we're connected to ordered by how likely they are to be on a seperate network.

fn connect(&self, our_conection_info: OurConnectionInfo, their_connection_info: TheirConnectionInfo)

Opens a connection to a remote peer. public_endpoint is the endpoint of the remote peer. udp_socket is a socket whose public address will be used by the other peer.

A rendezvous connection setup is different to the traditional BSD socket setup in which there is no client or server side. Both ends create a socket and send somehow its public address to the other peer. Once both ends know each other address, both must call this function passing the socket which possess the address used by the other peer and passing the other peer's address.

Only UDP-based protocols are supported. This means that you must use a uTP endpoint or nothing will happen.

On success Event::OnConnect with connected Endpoint will be sent to the event channel. On failure, nothing is reported. Failed attempts are not notified back up to the caller. If the caller wants to know of a failed attempt, it must maintain a record of the attempt itself which times out if a corresponding Event::OnConnect isn't received. See also Process for Connecting for details on handling of connect in different protocols.

fn get_known_external_endpoints(&self) -> Vec<Endpoint>

Get already known external endpoints without any upnp mapping

fn prepare_connection_info(&mut self, result_token: u32)

Lookup a mapped udp socket based on result_token