Struct crust::Service

pub struct Service {
    // some fields omitted
}

A structure representing a connection manager.

This abstraction has a hidden dependency on a config file. Refer to the docs for FileHandler and an example config file flowchart for more information.

Methods

impl Service

fn new(event_tx: CrustEventSender, service_discovery_port: u16) -> Result<Service, Error>

Constructs a service. User needs to create an asynchronous channel, and provide the sender half to this method. Receiver will receive all Events from this library.

fn stop_bootstrap(&mut self)

Stop the bootstraping procedure

fn start_listening_tcp(&mut self) -> Result<()>

Starts accepting TCP connections.

fn start_listening_utp(&mut self) -> Result<()>

Starts accepting uTP connections.

fn start_service_discovery(&mut self)

Starts listening for beacon broadcasts.

fn get_ordered_helping_nodes(&self) -> Vec<SocketAddr>

Get the hole punch servers addresses of nodes that we're connected to ordered by how likely they are to be on a seperate network.

fn send(&self, id: &PeerId, data: Vec<u8>) -> Result<()>

Send the given data to the peer with the given PeerId.

fn disconnect(&self, id: &PeerId) -> bool

Disconnect from the given peer and returns whether there was a connection at all.

fn _connect(&self, our_connection_info: OurConnectionInfo, their_connection_info: TheirConnectionInfo)

Opens a connection to a remote peer. public_endpoint is the endpoint of the remote peer. udp_socket is a socket whose public address will be used by the other peer.

A rendezvous connection setup is different to the traditional BSD socket setup in which there is no client or server side. Both ends create a socket and send somehow its public address to the other peer. Once both ends know each other address, both must call this function passing the socket which possess the address used by the other peer and passing the other peer's address.

Only UDP-based protocols are supported. This means that you must use a uTP endpoint or nothing will happen.

On success Event::NewPeer with connected PeerId will be sent to the event channel. On failure, nothing is reported. Failed attempts are not notified back up to the caller. If the caller wants to know of a failed attempt, it must maintain a record of the attempt itself which times out if a corresponding Event::NewPeer isn't received. See also Process for Connecting for details on handling of connect in different protocols.

fn connect(&self, our_connection_info: OurConnectionInfo, their_connection_info: TheirConnectionInfo)

Connect, but only via TCP.

fn prepare_connection_info(&mut self, result_token: u32)

Lookup a mapped udp socket based on result_token

fn id(&self) -> PeerId

Returns our ID.